docs/dbus-broker: update man-page
Fill in a lot more information on the internal API the broker provides.
This is still not finished, but at least most of the interfaces are
mentioned now.

Note that this interface is not stable, yet. It is still subject to
change. Since the reference-launcher is the only consumer for now, there
is no intention to stabilize the API, yet.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>