launch/nss-cache: index by ID and NAME
The nss system can return more information than just resolving Names to
IDs. The passwd/group structures contain a lot more details about a
specific entry. Until now, we did not need it. However, the <user> XML
attribute will need access to pw_gid, which we do not provide so far.

Rather than special casing the <user> resolver, lets make sure the
NSS-Cache can return all information. This means, every entry we have
must cache the "struct passwd" / "struct group" entry. Unfortunately,
this first means we need to extend the NSS-Cache to support resolving
numeric IDs via getpwuid(3) / getgrgid(3), so proper cache entries are
available even if queried by numeric ID.

Following this, we extend the cache to include a copy of the "struct
passwd/group" inside of the cache-node. The entry is cleared to 0 and
only the fields we need are cached. This is important, since allocation
of these structures is OS specific, and trailing data is often used.
Hence, we cannot easily memcpy() them. Instead, we must allocate and
instantiate them manually.

Note that for simplicity, this patch makes the nss_cache_add() call the
backend for both the populate-call *AND* the lookup. This means, we
will do 2 tree-traversals for cache-misses during the lookup, but that
should be acceptable given that we then reduce code-duplication.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>