message: make sender_id available in MessageMetadata
The sender_id is required when broadcasting, in the future we would like to provide
the MessageMetadata directly instead of MatchFilter, so move the sender_id here.

Note that this is different from the sender field, which is provided by the sender,
whereas the sender_id is the authorotative id set by the driver.

Signed-off-by: Tom Gundersen <teg@jklm.no>