driver/monitor: fix a regression
When reworking the broadcast handing, the filtering of monitoring was
broken. Pass in the matadata from the message object, rather than an
uninitialized one.

Note that this still only works for messages being forwarded, messages
originating in the driver do not have their metadata initalized, however
this is not a regression, it never worked correctly. It will be fixed
in a follow-up patch.

Signed-off-by: Tom Gundersen <teg@jklm.no>