user: optionally log whenever a charge fails
Having a quota exceeded is a global failure between two UID's, rather than
a failure of a specific transaction. Therefore let's do the logging in one place
where the actual charge happens.

A log object is pinned when the user registry is initialized, which is then always
used. If NULL is passed at initialization, all log messages are silently dropped.

The consequence of the quota failure should (if necessary) be logged independently.

Signed-off-by: Tom Gundersen <teg@jklm.no>