launch: properly unescape the Exec directive from .service files
The current logic was not correctly unquoating strings, as shown
by issue #144.

To quote the dbus sources, for the code that parses Exec

```
The results are defined to be the same as those you would get from
a UNIX98 /bin/sh, as long as the input contains none of the
unsupported shell expansions. If the input does contain such
expansions, they are passed through literally.
```

The same semantics is provided by a glib helper function, so we can
simply use that.

Signed-off-by: Tom Gundersen <teg@jklm.no>