peer/queue: treat directed signals exactly like method calls with no-reply-expected
This means that matches only ever affect broadcasts and directed signals
that can't be delivered due to resource exhaustion are silently dropped,
just like method calls are if the NO_REPLY_EXPECTED flag is set.

The implication is probably that directed signals should be avoided
altogether in new software that ought to be robust.

Signed-off-by: Tom Gundersen <teg@jklm.no>