apparmor: add skeleton support
The patches necessary for AppArmor support are not upstream in the kernel,
so we also do not implement AppArmor support upstream in dbus-broker.

However, this makes us respect the configuration to the extent possible.
AppArmor can be in one of three modes:
  * disable: we do nothing.
  * enabled: if the kernel has AppArmor support we warn about the discrepancy and continue,
             otherwise this is the same as 'disabled'.
  * required: refuse to start (or stop on reload). This means the broker should not be running
              unless AppArmor is being enforced.

The broker gains a new boolean argumnet, which unconditionally enables or disables AppArmor
(so any conditionals on the kernel state should be checked in the launcher). For now this
argument must be 'false'.

Signed-off-by: Tom Gundersen <teg@jklm.no>