util/dirwatch: add directory watch
The configurations and service files of dbus-daemon are watched for
local modifications. In fact, dbus-daemon monitors the sourced
*directories* and triggers a reload on *any* WRITE (including
new/moved/removed files). This only affects sourced directories, but
not explicitly sourced files.

For compatibility reasons we have to support the same. Hence, add a
directory-watch to our utilities. It is designed to do exactly what
dbus-daemon requires, not more, not less.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>