driver: return correct introspection data for non-canonical object paths
Ideally the driver would only implement the object /org/freedesktop/DBus,
however, for backwards compatibility we must allow some methods to be
called on any object path. We will divert slightly from the reference
implementation in the way we do the intropsection though, and expose it
as if only the canonical path exists.

We must also expose the canonical path as a node on each of its parents,
which was not done correctly in the past.

Signed-off-by: Tom Gundersen <teg@jklm.no>