audit: retain CAP_AUDIT_WRITE in the ambient capability set when dropping caps
Since we are not running at root, all caps will be dropped on execve(), unless
they are also in the ambient capability set, being in the inheritable set is
not sufficient.

This ensures that dbus-broker retains CAP_AUDIT_WRITE (when enabled), and that
dbus-broker-launch still does not.

This fixes issue #159.

Signed-off-by: Tom Gundersen <teg@jklm.no>