launcher/limits: reinterpret the limits for the broker
Limits in dbus-daemon are per-connection, on top of which there is a
connections-per-user limit. As there is no privilege separation
between connections of the same user, from a security point-of-view,
what is important is how much resources a certain user can consume
in total accross all its peers.

In dbus-broker, we do all accounting per-user, exactly to capture this
property. This patch reinterprets the standard dbus limits, as per-user
limits, by simply multiplying the per-connection limit by the
connections-per-user limit.

This means that a given connection can now potentially consume more
resources under dbus-broker than under dbus-daemon, though the total
ammount of resources a malicious user can consume remains the same.

dbus-broker accounts all bytes consumed by a given user, regardless in
what way (including objects used for internal bookkeeping), rather than
having separate accounting for each type of object. The only exception
is that we have an explicit limit on the number of match rules, as their
behavior can be linear (rather than logarithmic, as everything else is).

We choose to use the 'max-outgoing-bytes' dbus-daemon limit as our 'bytes'
limit as that most closely matches our quota logic (we account the number
of bytes queued an a given user, and further split that into quotas per
source connection).

Currently, we do not support explicit limits on bytes and FDs per message,
but that could potentially be added in follow-up commits.

It is worth noting that due to the hierarchical limits of dbus-daemon,
the total per-user limits end up being a lot higher by default than what
is probably necessary (and possibly safe). We could with benefit reduce
the defaults, though the dbus-daemon would perhaps then struggle (as they
would either allow too few connecitons per user, or not sufficient
resources per connection).

Signed-off-by: Tom Gundersen <teg@jklm.no>