test: introduce util_broker_settle()
This is a simple new helper that waits till the broker has settled. It
simply sends a synchronous method call to guarantee the incoming queue
of the broker is fully handled.

For now the client is a oneshot object, so once the function returns the
broker might still be busy disconnecting the client. If that turns out
to be an issue, we can always keep the client as part of the Broker
object, and only create a new one if it is NULL.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>