peer/dispatch: log on forced disconnect
If a peer exhausts its own resources, or violates the DBus protocl,
it is forcibly disconnected. Log that this happens. In the case of
resource exhaustion, the precise cause (though not the effect) is
already logged in the user layer. However, we may still want to
add similar logging for the precise protocol violation that occurs
to the driver layer.

Signed-off-by: Tom Gundersen <teg@jklm.no>