connection: only send/receive FDs to peers who support it
If a peer does not negotiate support for file descriptor passing,
consider it a protocol violation for the peer to send FDs anyway.

Also, refuse sending FDs to the peer, and consider any attempts to
do so in the same way as resource exhaustion. That is, return an
error to the sender in case of method calls, or disconnect the
receiver in case of replies or broadcasts.

Signed-off-by: Tom Gundersen <teg@jklm.no>