peer/dispatch: parse message metadata before passing into the driver
This is not a functional change, as it simply moves the point at which
the metadata is parsed from driver.c to peer.c. However, we now know
that only messages with valid metadata are passed to driver.c, which
will allow us to improve the logging in a follow-up commit.

Signed-off-by: Tom Gundersen <teg@jklm.no>
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>