message/header: use an explicit error code for missing FDs
Single out missing auxillary FDs for sepcial logging. We require
at least as many FDs passed to us as is declared in the header,
though ensuring this may sometimes be out of the hands of the client
library, so let us make it easier to debug.

In particular, if the SELinux policy installed on the system is not
right, it may cause the kernel to drop auxillary filedescriptors
silently (messages are logged to the audit system, but dbus-broker
wouldn't know).

Signed-off-by: Tom Gundersen <teg@jklm.no>