units/system: run in the root network namespace
The SELinux netlink socket can only be created in the root net
namespace. libselinux does not support (as far as I can tell)
providing the socket as a .socket unit, so we must run the broker
in the root network namespace if we want to be notified about
SELinux policy changes.

This fixes issue #185.

Signed-off-by: Tom Gundersen <teg@jklm.no>