launch: allow running without a scope
The scope (system or user) determines which systemd instance to
connect to (the system or user instance) in order to start
dbus-activated services.

If no scope is specified we were defaulting to "system".

Change the logic to allow running without a scope at all and make
that the new default (all users specify one, as far as I know, so
this should not change the behavior of existing users).

When running without a scope, name activation is disabled. That
means that we no longer expose names as activatable (but warn
from the launcher in case such names are configured), and we ignore
calls to change the activation environment.

Signed-off-by: Tom Gundersen <teg@jklm.no>