launch: do not log on destructive transactions
If we try to activate a service that is in the process of being stopped,
systemd refuses with an error message like:

dbus-broker-launch[760]: Activation request for 'org.freedesktop.login1' failed: Transaction for systemd-logind.service/start is destructive (local-fs.target has 'stop' job queued, but 'start' is included in transaction).

This is expected behavior, especially on shutdown, so we should not be
logging about that.

Signed-off-by: Tom Gundersen <teg@jklm.no>