launch/activate: do not spam the logs with NoSuchUnit error messages
A supported configuration is to have an activatable name pointing to
a non-existent service. This is useful in case the service is optional,
and enabling it installs the needed alias.

This allows the admin to control whether or not a service should be
dbus activatable.

Signed-off-by: Tom Gundersen <teg@jklm.no>
(use more descriptive variable-name for state-tracking)
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>