log: add support for the journal catalog
The log_append_here() call is extended to take an optional
@id argument, which is a 128 bit number formatted as a hex
string. If the argument is present the MESSAG_ID field
is appended to the journal entry.

This allows additional information to be provided for log
messages.

journalctl will render the catalog entries when passed the
'-x' option. Like so:
```
Mar 28 20:33:34 teg-x270 dbus-broker-launch[751]: Activation request for 'org.freedesktop.resolve1' failed: The systemd unit 'dbus-org.freedesktop.resolve1.service' could not be found.
-- Subject: Activation request failed due to missing unit
-- Defined-By: dbus-broker
-- Support: https://groups.google.com/forum/#!forum/bus1-devel
--
-- systemd failed to activate the name org.freedesktop.resolve1
-- because the the service dbus-org.freedesktop.resolve1.service could not be
-- found.
--
-- A possible reason for this is that dbus-org.freedesktop.resolve1.service is an
-- alias to a disabled service. This is allowed by design, to give the
-- administrator a way to disable name activation for a given service while
-- still allowing it to be installed and potentially started in other ways.
--
-- In order not to flood the logs, this message is only emitted once per name.
```

Signed-off-by: Tom Gundersen <teg@jklm.no>