build: pull in c-stdaux
We are about to drop the experimental c-sundry dependency. It was our
initial solution to pull in all kinds of helpers that have no other
home. Now with 4 years passed we have a separate home for all of the
crucial C helpers. Anything left in c-sundry can now be pulled into
dbus-broker directly, at which point c-sundry is no longer needed.

Unlike dbus-broker, the other dependents of the c-util projects never
pulled in c-sundry. This worked fine in most cases. However, some of the
macros from c-macro.h had to be copied into each dependent. This was not
really sustainable, so c-stdaux is now around to provide auxiliary
macros and helpers around the different standard libraries.

We now pull in c-stdaux, replace all the c-macro.h users, and then
eventually drop c-sundry.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>