dbus-broker: switch to c-stdaux
This switches dbus-broker to use c-stdaux. This is our new dependency
for auxiliary helpers around standard libraries. It replaces most of
the c-sundry helpers.

Since c-sundry is being phased out, the remaining bits from c-sundry
are pulled into src/util/ in this commit. This is needed to get rid of
c-macro.h fully.

Note that c-sundry is dropped from meson.build, but remains a
submodule. A followup commit will make sure to drop the submodule.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>