dbus-broker: switch to c_assert()
Use c_assert() over normal assert() in all our non-test code. See the
documentation of c_assert() of the advantages. In short: It always
evaluates the condition (including side-effects), and relies on
dead-code elimination to drop assertions in the NDEBUG case. This fits
our use of assert() perfectly, so lets switch over and get rid of the
compiler warnings in the NDEBUG case.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>