launch/service: fix state tracking
The 'state' field is currently used for 2 purposes: To track whether a
service was pushed into the broker, and to track whether a service was
found during reload. While both can be tracked in the same state, it
makes things overly complex and the current solution is definitely
wrong. We do not consider failing configuration-reloads, which happen
to leave the reload-tag untouched. In those cases, we trigger the wrong
assertion in the activation path.

This patch splits the state into separate booleans. One to track
whether or not a services was pushed into the broker ('running'),
and one as throw-away tag for reload tracking ('reload_tag'). Note that
the value of the reload-tag is only correct *DURING RELOADS*. It is
garbage at all other times. Do not verify it. It is solely meant to be
used by the launcher to track services across reloads!

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>