launch: mark new services as loaded on configuration reload
When we reload the bus configuration, we must create a diff of all
services before and after the reload. We use a `reload_tag` to mark
services as "updated" when we find them during a reload. However, we
never marked services as "updated" when they were new during the
reload. This caused those services to be removed again at the end of
the configuration reload, since the launcher assumed those services
were not found during this reload-round.

Fix this and explicitly mark all added services during a reload.

This was a regression from:

        commit b32683eff39160232879ba899947cd456b19be9e
        Author: David Rheinsberg <david.rheinsberg@gmail.com>
        Date:   Wed Apr 17 09:13:07 2019 +0200

            launch/service: fix state tracking

..which itself was a regression fix for the launch/service split.

Reported-by: Adam Williamson <awilliam@redhat.com>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>