dbus/socket: treat MSG_CTRUNC gracefully
As it turns out, LSMs allow clients to trigger a MSG_CTRUNC on the
remote side of a unix socket. Whenever LSMs reject the transmission of
an FD, they will simply drop the FD and set MSG_CTRUNC, without any
other error notification.

Therefore, we must assume any occurance of MSG_CTRUNC is trigger by a
client. This makes it impossible to consider MSG_CTRUNC for any other
error handling, and as such we are left to disconnecting the client and
ignoring the flag.

Luckily, MSG_CTRUNC is expected for any other event, so we only used it
for diagnostics so far.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>