bus/driver: allow eavesdrop filters for monitors
As it turns out `dbus-monitor` unconditionally prepends `eavesdrop=true`
to all filters it installs. This was necessary back in the days when it
actually used eavesdropping, but with BecomeMonitor() this is no longer
necessary. However, the code is still there and used in the wild.

To support `dbus-monitor` with custom matches, this commit makes
BecomeMonitor() treat the matches special and allows parsing `eavesdrop`
rules just like we used to:

        commit ce67e4e7091416d8cd7021b63babbfed7a01d7fd
        Author: Tom Gundersen <teg@jklm.no>
        Date:   Thu Aug 3 08:40:52 2017 +0200

            match: drop support for eavesdropping

We still silently ignore the flag, since it carries no meaning for
monitors, and in all other cases we do not allow it, anyway.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>