broker: force SO_PEERSEC if new option linux-4-17 is set
This adds a new build option called 'linux-4-17' (default: false),
which allows to tell the build system that linux-v4.17 is a
hard-requirement and its features can be used and relied upon.

This allows us to then use SO_PEERSEC over reading /proc to get selinux
labels.

Note that the idea here is to smoothly transition to requiring new
kernels. We do not want to enforce this, yet. But we want to allow
people to make use of newer kernels in controlled environments.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>