broker/broker: use SO_PEERCRED.pid over getppid()
When querying for LSM credentials of our parent, we want the strategy
forward to be using SO_PEERSEC. In the fallback-path, we use /proc on
the parent pid. To reduce the differences between those paths, actually
use the pid from the controller-socket over getppid(). This now reduces
the differences to just where we fetch the data from.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>