util/sockopt: keep preallocation local and small
Keep the initialization of the pre-allocation size local to the
function that allocates. Otherwise, we have this random `n = 64`
somewhere at the start of the function, with no explanation why and
what it does.

Furthermore, keep the pre-allocation small. There is little reason to
keep 64 entries allocated for every user. There is no real harm in
querying the kernel twice during connect, in case users really do have
huge supplementary group arrays.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>