util/sockopt: sort and deduplicate gid-array
We want predictable behavior in our APIs, unless really not possible.
Therefore, let us prepare for exporting supplementary groups in the
DBus driver API and make sure the group arrays are sorted and
deduplicated. The kernel guarantees sorted arrays, but does not
deduplicate. But lets just sort ourselves. This is fast if the array is
already sorted.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>