launch/main: fix error propagation
We incorrectly forward launcher errors from the launcher namespace to
the main application namespace. This is wrong, these are separate
namespaces, so we have to fold the error.

This fixes an issue were we would treat LAUNCHER_E_INVALID_CONFIG as
MAIN_EXIT, and thus not return with an error code.

Reported-by: Jan Macku
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>