util: add systemd unit escape helpers
Add a new utility library that helps us with escaping systemd unit
names. This is the same functionality as `systemd-escape(1)`, but as a
code library. Unfortunately, no `sd-escape` is provided, so we have to
copy the logic and implement it ourselves.

This will be used by the launcher to escape unit-names generated from
dbus-names, before creating the transient units.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>