util/audit: Revert UTIL_AUDIT_TYPE_SELINUX_ERROR audit type.
After discussions on the SELinux mail list related to the content of audit
messages that libselinux generates[1][2], it was determined that no
SELINUX_ERROR messages are auditable after all.  In the future, if there
are ones that need to be audited, they will have a different value in the
libselinux log callback and this can be restored.

[1] bus1/dbus-broker#240
[2] https://lore.kernel.org/selinux/CAEjxPJ6=s504ZdWevGALxxksCF4tsQ46meD8iCg7sCFoV9sKSA@mail.gmail.com/

Signed-off-by: Chris PeBenito <chpebeni@linux.microsoft.com>