bus/match: flush monitor matches on disconnect
We did not correctly flush monitor matches on peer disconnect, but only
the normal subscription objects. Fix this and properly flush all matches
in a registry on disconnect.

Background: We distinguish match-owners and match-registries. The owner
collects all matches of a peer and drops them when the peer disconnects.
It is merely a management structure for matches of each individual peer.
The match-registry, however, is a slot collecting all the matches that
match a particular object. This gives us O(1) lookup of matches on every
message transaction, as long as we have a suitable registry object.
When a peer disconnects, we first delete all the matches that peer owns,
then we unlink all remote matches that are linked on the peer's
registry, because those matches will no longer yield any results with
that peer gone.
Furthermore, we split between monitor matches and normal subscription
matches, so monitor operations do not affect normal operation unless a
monitor is active. Unfortunately, we never correctly flushed the monitor
matches on a registry on disconnect, meaning a peer with an active
monitor match on it will cause an assertion fault when it disconnects:

  dbus-broker:
  ../src/bus/match.c:1016:
  match_registry_deinit:
    Assertion `c_assert_result && "c_rbtree_is_empty(&registry->monitor_tree)"' failed.

This can be reproduced by running a monitor with a sender-match on a
specific peer unique-ID. Then make that peer disconnect. Note that this
is quite an unlikely situation, as unique-ID sender-matches are not
very common for monitors. This is probably why we didn't catch this
earlier.

This bug was likely introduced when we split normal subscription matches
and monitor matches for faster monitor lookups (back in 2017). I didn't
fully trace this back, though.

Reported-by: Benjamin Berg
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>