dbus/socket: increase SIOCOUTQ grace range
The SIOCOUTQ ioctl returns the amount of pending outgoing data on a
socket, but as recently discovered might include temporary reference
counters because the kernel overloads this counter.

Luckily, individual SKB allocations on the outgoing queue are always
quite big (including `sizeof(struct sk_buff)`), thus we can safely
assume that if the counter is lower than 128 then the kernel buffer is
actually empty. At the same time, it gives us 128 extra references the
kernel can hold at any time without deadlocking the socket.

This is all an ugly workaround, but there is little we can do as long as
the kernel reports incorrect data.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>