launch/service: track systemd unit state
Track the `ActiveState` property of systemd units associated with dbus
services. This allows us to better track when a service activation fails
beyond just the `StartUnit()` call.

Until now we only ever tracked failure of the `StartUnit()` call itself.
However, this call does not wait for a service to be fully activated,
but merely waits for the start-job to be queued in systemd. This means,
we only see failures of the unit-verification and transaction checks,
but never get any results of the transaction itself.

In particular, this means that if any unit successfully spawns, but then
fails before claiming a bus-name, we never see this failure, because it
is reported only late by systemd.

This commit extends the launcher to track all systemd units we activate
for as long as a dbus-service is live. We only ever query the
`ActiveState` property, and thus can reliably track when a unit fails.
Note that we never track when a unit succeeds, because this information
is of no use in the launcher. Instead, we rely on the broker to
successfully track acquired names (which it already does). We only ever
tell the broker when a unit fails, and we now do that over the entire
lifetime of a dbus-service. This way, even late failures are reported
to the broker (which will ignore them, if they appeared after a service
already activated).

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>