util/log: keep maximum log size at 2 MiB
Lets reduce the maximum log-message size by a factor of 128. We never
ended up sending large dumps of logs (not even for debugging), given
that we can just dump information as bus messages. Furthermore, all our
logs operate in lossy mode so far, so this path wasn't used, anyway.

We still keep the code around, because an easy edit of the lossy mode
allows for better debugging during development. However, it currently
seems unlikely that this will end up in releases.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>