launch/service: drop fail-safe of transient units
Since the introduction of transient-unit support in the launcher, we
have passed `true` as last argument to `StartTransientUnit()`. This
boolean controls whether systemd considers `ExecStart=` failures as
fatal or not:

    commit 97fb4d80806db4ffd29d8821cb8d2f26cb5bc7ca
    Author: Tom Gundersen <teg@jklm.no>
    Date:   Sat May 13 13:17:33 2017 +0200

        launch/activate: start transient unit if no systemd unit is configured

It is hard to tell why we passed `true` initially, but it might be
because we hadn't configured any failure handling, yet. Regardless, we
never dropped this initial workaround.

This commit changes the transient unit handling to always consider
process failures as fatal.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>