launch/launcher: always subscribe to systemd events
The systemd bus API has a special guard to sending events on the bus.
Only if at least one client signals interest, it will send those events.
One does this via the Subscribe() call on systemd.

Note that the different systemd deamons (logind in particular) already
do this, so systemd will unconditionally send events on the system bus.
However, on the user bus, this is not necessarily the case. Hence, make
sure to always subscribe to those events.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>