launch/service: rely on systemd PropertiesChanged behavior
The systemd API guarantees that basic properties are always included
with value in the PropertiesChanged signals. They are never invalidated.
If that was the case, there would be no way to reliably track all state
changes of a unit.

Drop the explicit query behavior and simplify the property handling. We
can safely rely on this API behavior.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>