launch/service: track systemd jobs to close a race condition
Start tracking the `JobRemoved` signal of systemd to know the point at
which systemd fully handled our submitted job. There is no other way to
reliably know when we can start handling PropertiesChanged signals.
systemd does not guarantee any other barrier on the bus but the
JobRemoved signal.

This is unfortunate, since it requires us to watch *all* JobRemoved
signals of systemd. However, according to systemd developers, it is the
only supported way to track Job submission. Hence, we have to comply.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>