bus/peer: correctly install duplicate monitor matches
We currently import monitor-matches into a temporary MatchOwner so we
can report back errors before turning a peer into a monitor. However, we
incorrectly ignored duplicate matches when importing those temporary
matches into a peer. Fix this and properly import every user-instance of
a match.

Note that this is mostly a no-op, but we must consider the
name-references we acquire when linking matches to names. Without it, we
would lack suitable references for each match-instance.

Reported-by: David Bazile
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>