Restore compatibility with older libaudit/libselinux
The version requirements were bumped, but for example libselinux 3.2 is
available in few places (eg: not yet in Debian unstable).
But the new functionality is just 4 preprocessor defines, so it's very simple
and not intrusive to ifdef them instead. This allows backporting to older
LTS distros without having to patch dbus-broker.

Signed-off-by: Luca Boccassi <bluca@debian.org>